+++ src/DatasetBase/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add method `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::asArray` to access the items of the collection as fixed size array
 - Add method `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::reverse` to reverse the order of the collection items
 - Add method `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::isEmpty` to check whether the collection is empty without counting
+- Add aggregation methods `\Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\FloatCollection::sum`, `\Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\FloatCollection::max` and `\Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\FloatCollection::min` to reduce boilerplate code when aggregating a float collection
+- Add aggregation methods `\Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\IntegerCollection::sum`, `\Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\IntegerCollection::max` and `\Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\IntegerCollection::min` to reduce boilerplate code when aggregating an integer collection
 
 ### Changed